chore: Task 타입 파일 분리 #7
